+++ brokerio/__main__.py
@@ -1,7 +1,3 @@
-from eventlet import monkey_patch  # mandatory! leave at the very top
-
-monkey_patch()
-
 from .cli import main
 
 if __name__ == "__main__":