+++ brokerio/guard/Guard.py
@@ -15,19 +15,19 @@ class Guard:
     def __init__(self, url):
         self.url = url
         self.logger = init_logging("Guard", logging.DEBUG)
+        self.client = None
+        self.client_queue = None
 
     def run(self):
         self.logger.info("Start guard ...")
         ctx = mp.get_context('spawn')
-        client_queue = mp.Manager().Queue(12)
+        self.client_queue = mp.Manager().Queue(12)
         message_queue = mp.Manager().Queue(12)
-        client = ctx.Process(target=simple_client, args=("Guard",
-                                                         self.url, client_queue, message_queue))
-        client.start()
+        self.client = ctx.Process(target=simple_client, args=("Guard",
+                                                              self.url, self.client_queue, message_queue))
+        self.client.start()
 
+    def join(self):
         while True:
-            message = client_queue.get()
+            message = self.client_queue.get()
             self.logger.info("Guard received message: {}".format(message))
-
-            # while True:
-            #    new_input = input("Send string to broker as new task: ")